+++ lib/src/iter.rs
@@ -104,45 +104,3 @@ impl<'r> Iterator for IssueMessagesIter<'r> {
     }
 }
 
-
-/// Extension trait for everything that iterates over Strings, to remove comment lines
-/// (Lines starting with "#")
-pub trait WithoutComments<I, S>
-    where I: Iterator<Item = S> + Sized,
-          S: AsRef<str>
-{
-    fn without_comments(self) -> WithoutCommentsIter<I, S>;
-}
-
-/// Iterator type to be returned from WithoutComments::without_comments.
-pub struct WithoutCommentsIter<I, S>(I)
-    where I: Iterator<Item = S> + Sized,
-          S: AsRef<str>;
-
-impl<I, S> WithoutComments<I, S> for I
-    where I: Iterator<Item = S> + Sized,
-          S: AsRef<str>
-{
-    fn without_comments(self) -> WithoutCommentsIter<I, S> {
-        WithoutCommentsIter(self)
-    }
-}
-
-impl<I, S> Iterator for WithoutCommentsIter<I, S>
-    where I: Iterator<Item = S> + Sized,
-          S: AsRef<str>
-{
-    type Item = S;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        while let Some(next) = self.0.next() {
-            // we do not trim whitespace here, because of code blocks in the message which might
-            // have a "#" at the beginning
-            if !next.as_ref().starts_with("#") {
-                return Some(next)
-            }
-        }
-        None
-    }
-}
-
